refactor: move entry point to application into root of src
